Moved UI components into ui folder, ran prettier
