+++ src/components/ui/Card.vue
